Fix lead management system
The AI is implementing the plan to fix the lead management system. This includes standardizing the UI to use modals for all lead operations, simplifying the lead creation form by removing unnecessary fields and making certain fields optional, resolving database constraint issues that cause conversion failures, and fixing the lead-to-customer conversion process. The AI will also ensure all navigation works correctly and comprehensive testing is performed.
