+++ src/components/leads/LeadDetailsDialog.tsx
@@ -1,334 +1,346 @@
-
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/dialog';
 import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { PriceInput } from '@/components/ui/price-input';
-import { Label } from '@/components/ui/label';
-import { Textarea } from '@/components/ui/textarea';
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
 import { Badge } from '@/components/ui/badge';
-import { Calendar, Phone, Mail, User, Building, Clock, MapPin, Edit, Save, X } from 'lucide-react';
-import { useSupabaseMutation } from '@/hooks/useSupabaseQuery';
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
+import { Edit, User, Calendar, Mail, Phone, Users, Clock, ArrowRight } from 'lucide-react';
+import { format } from 'date-fns';
+import { LeadForm } from './LeadForm';
+import { useEventTypeConfigs } from '@/hooks/useEventTypeConfigs';
 import { supabase } from '@/integrations/supabase/client';
 import { useAuth } from '@/contexts/AuthContext';
 import { toast } from 'sonner';
-import { useEventTypeConfigs } from '@/hooks/useEventTypeConfigs';
-import { sanitizeInput, validateEmail, validatePhone, validateTextLength } from '@/utils/security';
+import { useNavigate } from 'react-router-dom';
 
 interface LeadDetailsDialogProps {
+  lead: any;
   open: boolean;
   onOpenChange: (open: boolean) => void;
-  lead: any;
   onUpdate: () => void;
+  isEditMode?: boolean;
 }
 
 export const LeadDetailsDialog: React.FC<LeadDetailsDialogProps> = ({
+  lead,
   open,
   onOpenChange,
-  lead,
-  onUpdate
+  onUpdate,
+  isEditMode = false
 }) => {
   const { currentTenant } = useAuth();
-  const { data: eventTypeConfigs } = useEventTypeConfigs();
-  const [isEditing, setIsEditing] = useState(false);
-  const [formData, setFormData] = useState(lead || {});
-
-  const updateLeadMutation = useSupabaseMutation(
-    async (updatedData: any) => {
-      const { data, error } = await supabase
-        .from('leads')
-        .update(updatedData)
-        .eq('id', lead.id)
-        .select()
-        .single();
-      
-      if (error) throw error;
-      return data;
-    },
-    {
-      successMessage: 'Lead updated successfully!',
-      onSuccess: () => {
-        setIsEditing(false);
-        onUpdate();
-      }
-    }
-  );
-
-  const handleSave = () => {
-    // Validate and sanitize inputs before saving
-    const errors: string[] = [];
-    
-    if (!formData.name?.trim()) {
-      errors.push('Name is required');
-    }
-    
-    if (formData.email && !validateEmail(formData.email)) {
-      errors.push('Please enter a valid email address');
-    }
-    
-    if (formData.phone && !validatePhone(formData.phone)) {
-      errors.push('Please enter a valid phone number');
-    }
-    
-    if (formData.notes && !validateTextLength(formData.notes, 2000)) {
-      errors.push('Notes are too long (max 2000 characters)');
-    }
-    
-    if (errors.length > 0) {
-      toast.error(errors.join(', '));
-      return;
-    }
-    
-    const updateData = {
-      ...formData,
-      name: sanitizeInput(formData.name),
-      email: formData.email ? sanitizeInput(formData.email) : null,
-      phone: formData.phone ? sanitizeInput(formData.phone) : null,
-      company: formData.company ? sanitizeInput(formData.company) : null,
-      notes: formData.notes ? sanitizeInput(formData.notes) : null,
-      event_date: formData.event_date || null,
-      estimated_guests: formData.estimated_guests ? parseInt(formData.estimated_guests) : null,
-      estimated_budget: formData.estimated_budget ? parseFloat(formData.estimated_budget) : null,
-    };
-    
-    updateLeadMutation.mutate(updateData);
-  };
+  const navigate = useNavigate();
+  const { data: eventTypeConfigs = [] } = useEventTypeConfigs();
+  const [isEditing, setIsEditing] = useState(isEditMode);
+  const [isConverting, setIsConverting] = useState(false);
 
   const getStatusColor = (status: string) => {
     switch (status) {
-      case 'new': return 'bg-blue-100 text-blue-800';
-      case 'contacted': return 'bg-yellow-100 text-yellow-800';
-      case 'qualified': return 'bg-green-100 text-green-800';
-      case 'quoted': return 'bg-purple-100 text-purple-800';
-      case 'converted': return 'bg-emerald-100 text-emerald-800';
-      case 'lost': return 'bg-red-100 text-red-800';
-      default: return 'bg-gray-100 text-gray-800';
+      case 'new': return 'secondary';
+      case 'contacted': return 'default';
+      case 'qualified': return 'outline';
+      case 'quoted': return 'outline';
+      case 'won': return 'default';
+      case 'lost': return 'destructive';
+      default: return 'secondary';
     }
   };
 
-  if (!lead) return null;
+  const convertToCustomer = async () => {
+    if (!lead || !currentTenant?.id) return;
+
+    setIsConverting(true);
+    try {
+      // Create customer record
+      const customerData = {
+        tenant_id: currentTenant.id,
+        name: lead.name.trim(),
+        email: lead.email?.trim() || null,
+        phone: lead.phone?.trim() || null,
+        customer_type: 'individual',
+        notes: lead.notes?.trim() || null,
+        lead_id: lead.id,
+        active: true,
+        customer_since: new Date().toISOString().split('T')[0],
+        marketing_consent: false,
+        vip_status: false,
+        preferred_contact_method: 'email'
+      };
+
+      const { data: customer, error: customerError } = await supabase
+        .from('customers')
+        .insert([customerData])
+        .select()
+        .single();
+
+      if (customerError) throw customerError;
+
+      // Update lead status to won
+      const { error: leadError } = await supabase
+        .from('leads')
+        .update({ 
+          status: 'won',
+          conversion_date: new Date().toISOString(),
+          updated_at: new Date().toISOString()
+        })
+        .eq('id', lead.id)
+        .eq('tenant_id', currentTenant.id);
+
+      if (leadError) throw leadError;
+
+      toast.success('Lead successfully converted to customer!');
+      onOpenChange(false);
+      onUpdate();
+      navigate(`/customers/${customer.id}`);
+    } catch (error: any) {
+      console.error('Error converting lead to customer:', error);
+      
+      if (error.message?.includes('duplicate')) {
+        toast.error('A customer with this information already exists.');
+      } else if (error.message?.includes('not-null')) {
+        toast.error('Missing required customer information.');
+      } else {
+        toast.error(`Failed to convert lead: ${error.message || 'Please try again.'}`);
+      }
+    } finally {
+      setIsConverting(false);
+    }
+  };
+
+  const eventConfig = eventTypeConfigs.find(config => config.event_type === lead.event_type);
+
+  if (!open) return null;
 
   return (
     <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent className="max-w-2xl max-h-[90vh] overflow-y-auto">
+      <DialogContent className="max-w-4xl max-h-[90vh] overflow-y-auto">
         <DialogHeader>
-          <DialogTitle className="flex items-center justify-between">
-            <span>Lead Details</span>
-            <div className="flex items-center gap-2">
-              <Badge className={getStatusColor(formData.status || lead.status)}>
-                {(formData.status || lead.status)?.charAt(0).toUpperCase() + (formData.status || lead.status)?.slice(1)}
-              </Badge>
-              {!isEditing ? (
-                <Button size="sm" onClick={() => setIsEditing(true)}>
-                  <Edit className="h-4 w-4 mr-1" />
+          <div className="flex items-center justify-between">
+            <div>
+              <DialogTitle className="text-2xl">{lead.name}</DialogTitle>
+              <div className="flex items-center gap-2 mt-2">
+                <Badge variant={getStatusColor(lead.status)}>
+                  {lead.status}
+                </Badge>
+                {eventConfig && (
+                  <Badge 
+                    variant="outline" 
+                    style={{ 
+                      backgroundColor: eventConfig.color + '20',
+                      borderColor: eventConfig.color,
+                      color: eventConfig.text_color 
+                    }}
+                  >
+                    {eventConfig.display_name}
+                  </Badge>
+                )}
+              </div>
+            </div>
+            <div className="flex gap-2">
+              {!isEditing && (
+                <Button 
+                  variant="outline" 
+                  onClick={() => setIsEditing(true)}
+                >
+                  <Edit className="h-4 w-4 mr-2" />
                   Edit
                 </Button>
-              ) : (
-                <div className="flex gap-2">
-                  <Button size="sm" onClick={handleSave} disabled={updateLeadMutation.isPending}>
-                    <Save className="h-4 w-4 mr-1" />
-                    Save
-                  </Button>
-                  <Button size="sm" variant="outline" onClick={() => {
-                    setIsEditing(false);
-                    setFormData(lead);
-                  }}>
-                    <X className="h-4 w-4 mr-1" />
-                    Cancel
-                  </Button>
-                </div>
+              )}
+              {lead.status !== 'won' && !isEditing && (
+                <Button 
+                  onClick={convertToCustomer}
+                  disabled={isConverting}
+                  className="bg-primary hover:bg-primary/90"
+                >
+                  <ArrowRight className="h-4 w-4 mr-2" />
+                  {isConverting ? 'Converting...' : 'Convert to Customer'}
+                </Button>
+              )}
+              {isEditing && (
+                <Button 
+                  variant="outline" 
+                  onClick={() => setIsEditing(false)}
+                >
+                  Cancel
+                </Button>
               )}
             </div>
-          </DialogTitle>
+          </div>
         </DialogHeader>
 
-        <div className="space-y-6">
-          {/* Contact Information */}
-          <div className="space-y-4">
-            <h3 className="text-lg font-semibold flex items-center gap-2">
-              <User className="h-5 w-5" />
-              Contact Information
-            </h3>
-            
-            <div className="grid grid-cols-2 gap-4">
-              <div className="space-y-2">
-                <Label htmlFor="name">Name *</Label>
-                <Input
-                  id="name"
-                  value={formData.name || ''}
-                  onChange={(e) => setFormData({...formData, name: e.target.value})}
-                  disabled={!isEditing}
-                />
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="email">Email</Label>
-                <Input
-                  id="email"
-                  type="email"
-                  value={formData.email || ''}
-                  onChange={(e) => setFormData({...formData, email: e.target.value})}
-                  disabled={!isEditing}
-                />
-              </div>
-            </div>
-            
-            <div className="grid grid-cols-2 gap-4">
-              <div className="space-y-2">
-                <Label htmlFor="phone">Phone</Label>
-                <Input
-                  id="phone"
-                  value={formData.phone || ''}
-                  onChange={(e) => setFormData({...formData, phone: e.target.value})}
-                  disabled={!isEditing}
-                />
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="company">Company</Label>
-                <Input
-                  id="company"
-                  value={formData.company || ''}
-                  onChange={(e) => setFormData({...formData, company: e.target.value})}
-                  disabled={!isEditing}
-                />
-              </div>
-            </div>
-          </div>
+        <div className="mt-6">
+          {isEditing ? (
+            <LeadForm 
+              leadData={lead}
+              isEdit={true}
+              onSuccess={() => {
+                setIsEditing(false);
+                onUpdate();
+                toast.success('Lead updated successfully!');
+              }}
+              eventTypeConfigs={eventTypeConfigs}
+            />
+          ) : (
+            <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+              {/* Contact Information */}
+              <Card>
+                <CardHeader>
+                  <CardTitle className="flex items-center gap-2">
+                    <User className="h-5 w-5" />
+                    Contact Information
+                  </CardTitle>
+                </CardHeader>
+                <CardContent className="space-y-4">
+                  <div>
+                    <label className="text-sm font-medium text-muted-foreground">Full Name</label>
+                    <p className="text-foreground">{lead.name}</p>
+                  </div>
+                  {lead.email && (
+                    <div>
+                      <label className="text-sm font-medium text-muted-foreground">Email</label>
+                      <div className="flex items-center gap-2">
+                        <Mail className="h-4 w-4" />
+                        <a href={`mailto:${lead.email}`} className="text-primary hover:underline">
+                          {lead.email}
+                        </a>
+                      </div>
+                    </div>
+                  )}
+                  {lead.phone && (
+                    <div>
+                      <label className="text-sm font-medium text-muted-foreground">Phone</label>
+                      <div className="flex items-center gap-2">
+                        <Phone className="h-4 w-4" />
+                        <a href={`tel:${lead.phone}`} className="text-primary hover:underline">
+                          {lead.phone}
+                        </a>
+                      </div>
+                    </div>
+                  )}
+                  <div>
+                    <label className="text-sm font-medium text-muted-foreground">Source</label>
+                    <p className="text-foreground capitalize">{lead.source}</p>
+                  </div>
+                </CardContent>
+              </Card>
 
-          {/* Event Details */}
-          <div className="space-y-4">
-            <h3 className="text-lg font-semibold flex items-center gap-2">
-              <Calendar className="h-5 w-5" />
-              Event Details
-            </h3>
-            
-            <div className="grid grid-cols-2 gap-4">
-              <div className="space-y-2">
-                <Label htmlFor="event_type">Event Type</Label>
-                <Select
-                  value={formData.event_type || 'none'}
-                  onValueChange={(value) => setFormData({...formData, event_type: value === 'none' ? null : value})}
-                  disabled={!isEditing}
-                >
-                  <SelectTrigger>
-                    <SelectValue placeholder="Select event type" />
-                  </SelectTrigger>
-                    <SelectContent>
-                      <SelectItem value="none">No event type selected</SelectItem>
-                      {eventTypeConfigs?.map(config => (
-                        <SelectItem key={config.id} value={config.event_type}>
-                          {config.display_name}
-                        </SelectItem>
-                      ))}
-                      <SelectItem value="other">Other</SelectItem>
-                    </SelectContent>
-                </Select>
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="event_date">Event Date</Label>
-                <Input
-                  id="event_date"
-                  type="date"
-                  value={formData.event_date || ''}
-                  onChange={(e) => setFormData({...formData, event_date: e.target.value})}
-                  disabled={!isEditing}
-                />
-              </div>
-            </div>
-            
-            <div className="grid grid-cols-2 gap-4">
-              <div className="space-y-2">
-                <Label htmlFor="estimated_guests">Estimated Guests</Label>
-                <Input
-                  id="estimated_guests"
-                  type="number"
-                  value={formData.estimated_guests || ''}
-                  onChange={(e) => setFormData({...formData, estimated_guests: e.target.value})}
-                  disabled={!isEditing}
-                />
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="estimated_budget">Estimated Budget (£)</Label>
-                <PriceInput
-                  value={parseFloat(formData.estimated_budget) || 0}
-                  onChange={(value) => setFormData({...formData, estimated_budget: value.toString()})}
-                  disabled={!isEditing}
-                />
-              </div>
-            </div>
-          </div>
+              {/* Event Details */}
+              <Card>
+                <CardHeader>
+                  <CardTitle className="flex items-center gap-2">
+                    <Calendar className="h-5 w-5" />
+                    Event Details
+                  </CardTitle>
+                </CardHeader>
+                <CardContent className="space-y-4">
+                  <div>
+                    <label className="text-sm font-medium text-muted-foreground">Event Type</label>
+                    <p className="text-foreground">{eventConfig?.display_name || lead.event_type || 'Not specified'}</p>
+                  </div>
+                  {lead.date_of_interest && (
+                    <div>
+                      <label className="text-sm font-medium text-muted-foreground">Date of Interest</label>
+                      <p className="text-foreground">
+                        {format(new Date(lead.date_of_interest), 'EEEE, dd MMMM yyyy')}
+                      </p>
+                    </div>
+                  )}
+                  <div>
+                    <label className="text-sm font-medium text-muted-foreground">Guest Information</label>
+                    <div className="flex items-center gap-2">
+                      <Users className="h-4 w-4" />
+                      <span>{lead.men_count + lead.ladies_count} guests ({lead.guest_mixture})</span>
+                    </div>
+                    <p className="text-sm text-muted-foreground mt-1">
+                      {lead.men_count} men, {lead.ladies_count} ladies
+                    </p>
+                  </div>
+                  {lead.estimated_budget && (
+                    <div>
+                      <label className="text-sm font-medium text-muted-foreground">Estimated Budget</label>
+                      <p className="text-foreground">£{parseFloat(lead.estimated_budget).toLocaleString()}</p>
+                    </div>
+                  )}
+                </CardContent>
+              </Card>
 
-          {/* Lead Management */}
-          <div className="space-y-4">
-            <h3 className="text-lg font-semibold">Lead Management</h3>
-            
-            <div className="grid grid-cols-2 gap-4">
-              <div className="space-y-2">
-                <Label htmlFor="status">Status</Label>
-                <Select
-                  value={formData.status || 'new'}
-                  onValueChange={(value) => setFormData({...formData, status: value})}
-                  disabled={!isEditing}
-                >
-                  <SelectTrigger>
-                    <SelectValue />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="new">New</SelectItem>
-                    <SelectItem value="contacted">Contacted</SelectItem>
-                    <SelectItem value="qualified">Qualified</SelectItem>
-                    <SelectItem value="quoted">Quoted</SelectItem>
-                    <SelectItem value="converted">Converted</SelectItem>
-                    <SelectItem value="lost">Lost</SelectItem>
-                  </SelectContent>
-                </Select>
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="source">Source</Label>
-                <Select
-                  value={formData.source || 'none'}
-                  onValueChange={(value) => setFormData({...formData, source: value === 'none' ? null : value})}
-                  disabled={!isEditing}
-                >
-                  <SelectTrigger>
-                    <SelectValue />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="none">No source selected</SelectItem>
-                    <SelectItem value="website">Website</SelectItem>
-                    <SelectItem value="social_media">Social Media</SelectItem>
-                    <SelectItem value="referral">Referral</SelectItem>
-                    <SelectItem value="google">Google</SelectItem>
-                    <SelectItem value="phone">Phone</SelectItem>
-                    <SelectItem value="walk_in">Walk In</SelectItem>
-                    <SelectItem value="other">Other</SelectItem>
-                  </SelectContent>
-                </Select>
-              </div>
-            </div>
-            
-            <div className="space-y-2">
-              <Label htmlFor="notes">Notes</Label>
-              <Textarea
-                id="notes"
-                value={formData.notes || ''}
-                onChange={(e) => setFormData({...formData, notes: e.target.value})}
-                disabled={!isEditing}
-                rows={4}
-              />
-            </div>
-          </div>
+              {/* Appointment Information */}
+              <Card>
+                <CardHeader>
+                  <CardTitle className="flex items-center gap-2">
+                    <Clock className="h-5 w-5" />
+                    Appointment Information
+                  </CardTitle>
+                </CardHeader>
+                <CardContent className="space-y-4">
+                  {lead.appointment_date ? (
+                    <div>
+                      <label className="text-sm font-medium text-muted-foreground">Scheduled Appointment</label>
+                      <p className="text-foreground">
+                        {format(new Date(lead.appointment_date), 'EEEE, dd MMMM yyyy')}
+                      </p>
+                    </div>
+                  ) : (
+                    <div>
+                      <p className="text-muted-foreground">No appointment scheduled</p>
+                    </div>
+                  )}
+                  {lead.date_of_contact && (
+                    <div>
+                      <label className="text-sm font-medium text-muted-foreground">Date of Contact</label>
+                      <p className="text-foreground">
+                        {format(new Date(lead.date_of_contact), 'dd MMMM yyyy')}
+                      </p>
+                    </div>
+                  )}
+                </CardContent>
+              </Card>
 
-          {/* Timestamps */}
-          <div className="space-y-2 text-sm text-gray-600">
-            <p>Created: {new Date(lead.created_at).toLocaleDateString()}</p>
-            {lead.updated_at && (
-              <p>Last Updated: {new Date(lead.updated_at).toLocaleDateString()}</p>
-            )}
-            {lead.last_contacted_at && (
-              <p>Last Contacted: {new Date(lead.last_contacted_at).toLocaleDateString()}</p>
-            )}
-          </div>
+              {/* Timeline */}
+              <Card>
+                <CardHeader>
+                  <CardTitle>Timeline</CardTitle>
+                </CardHeader>
+                <CardContent className="space-y-4">
+                  <div className="space-y-3">
+                    <div className="flex justify-between items-center">
+                      <span className="text-sm font-medium">Created</span>
+                      <span className="text-sm text-muted-foreground">
+                        {format(new Date(lead.created_at), 'dd/MM/yyyy HH:mm')}
+                      </span>
+                    </div>
+                    {lead.updated_at && lead.updated_at !== lead.created_at && (
+                      <div className="flex justify-between items-center">
+                        <span className="text-sm font-medium">Last Updated</span>
+                        <span className="text-sm text-muted-foreground">
+                          {format(new Date(lead.updated_at), 'dd/MM/yyyy HH:mm')}
+                        </span>
+                      </div>
+                    )}
+                    {lead.conversion_date && (
+                      <div className="flex justify-between items-center">
+                        <span className="text-sm font-medium">Converted to Customer</span>
+                        <span className="text-sm text-muted-foreground">
+                          {format(new Date(lead.conversion_date), 'dd/MM/yyyy HH:mm')}
+                        </span>
+                      </div>
+                    )}
+                  </div>
+                </CardContent>
+              </Card>
+            </div>
+          )}
+
+          {/* Notes */}
+          {!isEditing && lead.notes && (
+            <Card className="mt-6">
+              <CardHeader>
+                <CardTitle>Notes</CardTitle>
+              </CardHeader>
+              <CardContent>
+                <p className="text-foreground whitespace-pre-wrap">{lead.notes}</p>
+              </CardContent>
+            </Card>
+          )}
         </div>
       </DialogContent>
     </Dialog>
