--- src/pages/LeadEdit.tsx
@@ -1,340 +1,0 @@
-
-import React, { useState, useEffect } from 'react';
-import { ArrowLeft, Save, X } from 'lucide-react';
-import { Button } from '@/components/ui/button';
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { Textarea } from '@/components/ui/textarea';
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
-import { useSupabaseQuery, useSupabaseMutation } from '@/hooks/useSupabaseQuery';
-import { supabase } from '@/integrations/supabase/client';
-import { useAuth } from '@/contexts/AuthContext';
-import { useParams, useNavigate } from 'react-router-dom';
-import { toast } from 'sonner';
-
-import { useEventTypeConfigs } from '@/hooks/useEventTypeConfigs';
-
-export const LeadEdit = () => {
-  const { currentTenant } = useAuth();
-  const { data: eventTypeConfigs } = useEventTypeConfigs();
-  const { leadId } = useParams<{ leadId: string }>();
-  const navigate = useNavigate();
-  
-  const [formData, setFormData] = useState({
-    name: '',
-    email: '',
-    phone: '',
-    company: '',
-    event_type: '',
-    event_date: '',
-    estimated_guests: '',
-    estimated_budget: '',
-    source: 'website',
-    notes: '',
-    status: 'new'
-  });
-
-  const { data: lead, isLoading } = useSupabaseQuery(
-    ['lead', leadId],
-    async () => {
-      if (!currentTenant?.id || !leadId) return null;
-      
-      const { data, error } = await supabase
-        .from('leads')
-        .select('*')
-        .eq('id', leadId)
-        .eq('tenant_id', currentTenant.id)
-        .single();
-      
-      if (error) throw error;
-      return data;
-    }
-  );
-
-  useEffect(() => {
-    if (lead) {
-      setFormData({
-        name: lead.name || '',
-        email: lead.email || '',
-        phone: lead.phone || '',
-        company: lead.company || '',
-        event_type: lead.event_type || '',
-        event_date: lead.event_date || '',
-        estimated_guests: lead.estimated_guests ? lead.estimated_guests.toString() : '',
-        estimated_budget: lead.estimated_budget ? lead.estimated_budget.toString() : '',
-        source: lead.source || 'website',
-        notes: lead.notes || '',
-        status: lead.status || 'new'
-      });
-    }
-  }, [lead]);
-
-  const updateLeadMutation = useSupabaseMutation(
-    async (updatedData: any) => {
-      const { data, error } = await supabase
-        .from('leads')
-        .update(updatedData)
-        .eq('id', leadId)
-        .select()
-        .single();
-      
-      if (error) throw error;
-      return data;
-    },
-    {
-      successMessage: 'Lead updated successfully!',
-      onSuccess: () => {
-        navigate(`/leads/${leadId}/view`);
-      }
-    }
-  );
-
-  const handleSave = () => {
-    const updateData = {
-      ...formData,
-      event_date: formData.event_date || null,
-      estimated_guests: formData.estimated_guests ? parseInt(formData.estimated_guests) : null,
-      estimated_budget: formData.estimated_budget ? parseFloat(formData.estimated_budget) : null,
-      email: formData.email || null,
-      phone: formData.phone || null,
-      company: formData.company || null,
-      event_type: formData.event_type || null,
-      notes: formData.notes || null,
-    };
-    
-    updateLeadMutation.mutate(updateData);
-  };
-
-  if (isLoading) {
-    return (
-      <div className="p-8 bg-gray-50 min-h-screen">
-        <div className="animate-pulse">
-          <div className="h-8 bg-gray-200 rounded w-1/4 mb-4"></div>
-          <div className="h-4 bg-gray-200 rounded w-1/2 mb-8"></div>
-          <div className="space-y-4">
-            <div className="h-32 bg-gray-200 rounded"></div>
-            <div className="h-32 bg-gray-200 rounded"></div>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  if (!lead) {
-    return (
-      <div className="p-8 bg-gray-50 min-h-screen">
-        <div className="text-center py-12">
-          <h2 className="text-2xl font-bold text-gray-900 mb-2">Lead Not Found</h2>
-          <p className="text-gray-600 mb-4">The lead you're looking for doesn't exist.</p>
-          <Button onClick={() => navigate('/leads')}>
-            <ArrowLeft className="h-4 w-4 mr-2" />
-            Back to Leads
-          </Button>
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="p-8 bg-gray-50 min-h-screen">
-      <div className="flex justify-between items-center mb-8">
-        <div className="flex items-center gap-4">
-          <Button variant="outline" onClick={() => navigate(`/leads/${leadId}/view`)}>
-            <ArrowLeft className="h-4 w-4 mr-2" />
-            Back to View
-          </Button>
-          <div>
-            <h1 className="text-3xl font-bold text-gray-900 mb-2">Edit Lead</h1>
-            <p className="text-gray-600">Update lead information and details</p>
-          </div>
-        </div>
-        
-        <div className="flex items-center gap-3">
-          <Button variant="outline" onClick={() => navigate(`/leads/${leadId}/view`)}>
-            <X className="h-4 w-4 mr-2" />
-            Cancel
-          </Button>
-          <Button onClick={handleSave} disabled={!formData.name || updateLeadMutation.isPending}>
-            <Save className="h-4 w-4 mr-2" />
-            {updateLeadMutation.isPending ? 'Saving...' : 'Save Changes'}
-          </Button>
-        </div>
-      </div>
-
-      <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-        {/* Contact Information */}
-        <Card>
-          <CardHeader>
-            <CardTitle>Contact Information</CardTitle>
-          </CardHeader>
-          <CardContent className="space-y-4">
-            <div className="grid grid-cols-2 gap-4">
-              <div className="space-y-2">
-                <Label htmlFor="name">Name *</Label>
-                <Input
-                  id="name"
-                  value={formData.name}
-                  onChange={(e) => setFormData({...formData, name: e.target.value})}
-                  required
-                />
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="company">Company</Label>
-                <Input
-                  id="company"
-                  value={formData.company}
-                  onChange={(e) => setFormData({...formData, company: e.target.value})}
-                />
-              </div>
-            </div>
-            
-            <div className="grid grid-cols-2 gap-4">
-              <div className="space-y-2">
-                <Label htmlFor="email">Email</Label>
-                <Input
-                  id="email"
-                  type="email"
-                  value={formData.email}
-                  onChange={(e) => setFormData({...formData, email: e.target.value})}
-                />
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="phone">Phone</Label>
-                <Input
-                  id="phone"
-                  value={formData.phone}
-                  onChange={(e) => setFormData({...formData, phone: e.target.value})}
-                />
-              </div>
-            </div>
-          </CardContent>
-        </Card>
-
-        {/* Event Details */}
-        <Card>
-          <CardHeader>
-            <CardTitle>Event Details</CardTitle>
-          </CardHeader>
-          <CardContent className="space-y-4">
-            <div className="grid grid-cols-2 gap-4">
-              <div className="space-y-2">
-                <Label htmlFor="event_type">Event Type</Label>
-                <Select value={formData.event_type || 'none'} onValueChange={(value) => setFormData({...formData, event_type: value === 'none' ? '' : value})}>
-                  <SelectTrigger>
-                    <SelectValue placeholder="Select event type" />
-                  </SelectTrigger>
-                    <SelectContent>
-                      <SelectItem value="none">None</SelectItem>
-                      {eventTypeConfigs?.map(config => (
-                        <SelectItem key={config.id} value={config.event_type}>
-                          {config.display_name}
-                        </SelectItem>
-                      ))}
-                      <SelectItem value="other">Other</SelectItem>
-                    </SelectContent>
-                </Select>
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="event_date">Event Date</Label>
-                <Input
-                  id="event_date"
-                  type="date"
-                  value={formData.event_date}
-                  onChange={(e) => setFormData({...formData, event_date: e.target.value})}
-                />
-              </div>
-            </div>
-            
-            <div className="grid grid-cols-2 gap-4">
-              <div className="space-y-2">
-                <Label htmlFor="estimated_guests">Estimated Guests</Label>
-                <Input
-                  id="estimated_guests"
-                  type="number"
-                  min="0"
-                  value={formData.estimated_guests}
-                  onChange={(e) => setFormData({...formData, estimated_guests: e.target.value})}
-                />
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="estimated_budget">Estimated Budget (£)</Label>
-                <Input
-                  id="estimated_budget"
-                  type="number"
-                  step="0.01"
-                  min="0"
-                  value={formData.estimated_budget}
-                  onChange={(e) => setFormData({...formData, estimated_budget: e.target.value})}
-                />
-              </div>
-            </div>
-          </CardContent>
-        </Card>
-
-        {/* Lead Management */}
-        <Card>
-          <CardHeader>
-            <CardTitle>Lead Management</CardTitle>
-          </CardHeader>
-          <CardContent className="space-y-4">
-            <div className="grid grid-cols-2 gap-4">
-              <div className="space-y-2">
-                <Label htmlFor="status">Status</Label>
-                <Select value={formData.status} onValueChange={(value) => setFormData({...formData, status: value})}>
-                  <SelectTrigger>
-                    <SelectValue />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="new">New</SelectItem>
-                    <SelectItem value="contacted">Contacted</SelectItem>
-                    <SelectItem value="qualified">Qualified</SelectItem>
-                    <SelectItem value="quoted">Quoted</SelectItem>
-                    <SelectItem value="converted">Converted</SelectItem>
-                    <SelectItem value="lost">Lost</SelectItem>
-                  </SelectContent>
-                </Select>
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="source">Source</Label>
-                <Select value={formData.source} onValueChange={(value) => setFormData({...formData, source: value})}>
-                  <SelectTrigger>
-                    <SelectValue />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="website">Website</SelectItem>
-                    <SelectItem value="social_media">Social Media</SelectItem>
-                    <SelectItem value="referral">Referral</SelectItem>
-                    <SelectItem value="google">Google</SelectItem>
-                    <SelectItem value="phone">Phone</SelectItem>
-                    <SelectItem value="walk_in">Walk In</SelectItem>
-                    <SelectItem value="other">Other</SelectItem>
-                  </SelectContent>
-                </Select>
-              </div>
-            </div>
-          </CardContent>
-        </Card>
-
-        {/* Notes */}
-        <Card>
-          <CardHeader>
-            <CardTitle>Notes</CardTitle>
-          </CardHeader>
-          <CardContent>
-            <div className="space-y-2">
-              <Label htmlFor="notes">Additional Notes</Label>
-              <Textarea
-                id="notes"
-                value={formData.notes}
-                onChange={(e) => setFormData({...formData, notes: e.target.value})}
-                placeholder="Any additional notes about this lead..."
-                rows={6}
-              />
-            </div>
-          </CardContent>
-        </Card>
-      </div>
-    </div>
-  );
-};
